Move transaction stuff somewhere, you know, more sane
