+++ src/pages/accounts/AccountView.vue
@@ -8,8 +8,8 @@ import Fuse from "fuse.js";
 import List from "../../components/List.vue";
 import Modal from "../../components/Modal.vue";
 import SearchBar from "../../components/SearchBar.vue";
-import TransactionEdit from "../../components/transactions/TransactionEdit.vue";
-import TransactionListItem from "../../components/transactions/TransactionListItem.vue";
+import TransactionEdit from "../transactions/TransactionEdit.vue";
+import TransactionListItem from "../transactions/TransactionListItem.vue";
 import { dinero, isNegative as isDineroNegative } from "dinero.js";
 import { intlFormat } from "../../transformers";
 import { ref, computed, toRefs, watch } from "vue";
