+++ src/pages/attachments/FileView.vue
@@ -5,7 +5,7 @@ import ActionButton from "../../components/buttons/ActionButton.vue";
 import DownloadButton from "../../components/buttons/DownloadButton.vue";
 import ErrorNotice from "../../components/ErrorNotice.vue";
 import List from "../../components/List.vue";
-import TransactionListItem from "../../components/transactions/TransactionListItem.vue";
+import TransactionListItem from "../transactions/TransactionListItem.vue";
 import TrashIcon from "../../icons/Trash.vue";
 import { ref, computed, watch, toRefs } from "vue";
 import { toTimestamp } from "../../transformers";
