+++ src/pages/transactions/TransactionEdit.vue
@@ -3,15 +3,15 @@ import type { Account } from "../../model/Account";
 import type { Location, LocationRecordParams } from "../../model/Location";
 import type { PropType } from "vue";
 import type { TransactionRecordParams } from "../../model/Transaction";
-import ActionButton from "../buttons/ActionButton.vue";
-import Checkbox from "../inputs/Checkbox.vue";
+import ActionButton from "../../components/buttons/ActionButton.vue";
+import Checkbox from "../../components/inputs/Checkbox.vue";
 import CheckmarkIcon from "../../icons/Checkmark.vue";
 import ConfirmDestroyTransaction from "./ConfirmDestroyTransaction.vue";
-import CurrencyInput from "../inputs/CurrencyInput.vue";
-import DateTimeInput from "../inputs/DateTimeInput.vue";
-import LocationField from "../../pages/locations/LocationField.vue";
-import TextAreaField from "../inputs/TextAreaField.vue";
-import TextField from "../inputs/TextField.vue";
+import CurrencyInput from "../../components/inputs/CurrencyInput.vue";
+import DateTimeInput from "../../components/inputs/DateTimeInput.vue";
+import LocationField from "../locations/LocationField.vue";
+import TextAreaField from "../../components/inputs/TextAreaField.vue";
+import TextField from "../../components/inputs/TextField.vue";
 import TrashIcon from "../../icons/Trash.vue";
 import { dinero, isNegative, isZero, toSnapshot } from "dinero.js";
 import { ref, computed, toRefs, onMounted } from "vue";
