+++ src/pages/transactions/TransactionView.vue
@@ -3,17 +3,17 @@ import type { Tag as TagObject, TagRecordParams } from "../../model/Tag";
 import type { Attachment } from "../../model/Attachment";
 import type { Location } from "../../model/Location";
 import type { Transaction } from "../../model/Transaction";
-import ConfirmDestroyFile from "../../pages/attachments/ConfirmDestroyFile.vue";
-import EditButton from "../buttons/EditButton.vue";
-import FileInput from "../../pages/attachments/FileInput.vue";
-import FileListItem from "../../pages/attachments/FileListItem.vue";
-import FileReattach from "../../pages/attachments/FileReattach.vue";
-import List from "../List.vue";
+import ConfirmDestroyFile from "../attachments/ConfirmDestroyFile.vue";
+import EditButton from "../../components/buttons/EditButton.vue";
+import FileInput from "../attachments/FileInput.vue";
+import FileListItem from "../attachments/FileListItem.vue";
+import FileReattach from "../attachments/FileReattach.vue";
+import List from "../../components/List.vue";
 import LocationIcon from "../../icons/Location.vue";
-import LocationView from "../../pages/locations/LocationView.vue";
-import Modal from "../Modal.vue";
-import NavAction from "../NavAction.vue";
-import TagList from "../tags/TagList.vue";
+import LocationView from "../locations/LocationView.vue";
+import Modal from "../../components/Modal.vue";
+import NavAction from "../../components/NavAction.vue";
+import TagList from "../../components/tags/TagList.vue";
 import TransactionEdit from "./TransactionEdit.vue";
 import { ref, computed, toRefs } from "vue";
 import { intlFormat, toTimestamp } from "../../transformers";
